+++ scrapyd/eggstorage.py
@@ -5,6 +5,7 @@ from glob import escape, glob
 
 from zope.interface import implementer
 
+from scrapyd.exceptions import DirectoryTraversalError
 from scrapyd.interfaces import IEggStorage
 from scrapyd.utils import sorted_versions
 
@@ -16,11 +17,13 @@ class FilesystemEggStorage(object):
         self.basedir = config.get('eggs_dir', 'eggs')
 
     def put(self, eggfile, project, version):
-        eggpath = self._eggpath(project, version)
-        eggdir = os.path.dirname(eggpath)
-        if not os.path.exists(eggdir):
-            os.makedirs(eggdir)
-        with open(eggpath, 'wb') as f:
+        path = self._egg_path(project, version)
+
+        directory = os.path.dirname(path)
+        if not os.path.exists(directory):
+            os.makedirs(directory)
+
+        with open(path, 'wb') as f:
             shutil.copyfileobj(eggfile, f)
 
     def get(self, project, version=None):
@@ -29,14 +32,12 @@ class FilesystemEggStorage(object):
                 version = self.list(project)[-1]
             except IndexError:
                 return None, None
-        return version, open(self._eggpath(project, version), 'rb')
+        return version, open(self._egg_path(project, version), 'rb')
 
     def list(self, project):
-        versions = [
-            os.path.splitext(os.path.basename(path))[0]
-            for path in glob(os.path.join(self.basedir, escape(project), "*.egg"))
-        ]
-        return sorted_versions(versions)
+        return sorted_versions(
+            [os.path.splitext(os.path.basename(path))[0] for path in glob(self._get_path(escape(project), "*.egg"))]
+        )
 
     def list_projects(self):
         if os.path.exists(self.basedir):
@@ -45,12 +46,21 @@ class FilesystemEggStorage(object):
 
     def delete(self, project, version=None):
         if version is None:
-            shutil.rmtree(os.path.join(self.basedir, project))
+            shutil.rmtree(self._get_path(project))
         else:
-            os.remove(self._eggpath(project, version))
+            os.remove(self._egg_path(project, version))
             if not self.list(project):  # remove project if no versions left
                 self.delete(project)
 
-    def _eggpath(self, project, version):
+    def _egg_path(self, project, version):
         sanitized_version = re.sub(r'[^A-Za-z0-9_-]', '_', version)
-        return os.path.join(self.basedir, project, f"{sanitized_version}.egg")
+        return self._get_path(project, f"{sanitized_version}.egg")
+
+    def _get_path(self, project, *trusted):
+        resolvedir = os.path.realpath(self.basedir)
+        projectdir = os.path.realpath(os.path.join(resolvedir, project))
+
+        if os.path.commonprefix((projectdir, resolvedir)) != resolvedir:
+            raise DirectoryTraversalError(f"{projectdir} is not under the {self.basedir} ({resolvedir}) directory")
+
+        return os.path.join(projectdir, *trusted)